requirements.txt
requests==2.22.0
Pillow==7.2.0
